requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
docker==7.1.0
fastapi==0.115.12
filelock==3.18.0
greenlet==3.1.1
h11==0.14.0
identify==2.6.9
idna==3.10
iniconfig==2.1.0
Mako==1.3.9
MarkupSafe==3.0.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
pytest-alembic==0.11.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
ruff==0.11.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.30.0
